Tema 1 entrega

+++ EjerciciosEntregaT1/src/ejercicios_entrega/Ejercicio10.java
@@ -6,10 +6,10 @@ public class Ejercicio10 {
 
 	public static void main(String[] args) {
 		// Definimos la variable donde se almacenarán los metros
-		float metros;
+		double metros;
 		
 		// Definimos la variable donde se almacenarán los centimetros
-		float centimetros;
+		double centimetros;
 		
 		// Definimos la variable donde se almacenarán los centímetros sin decimales
 		int centimetrosSinDecimales;
@@ -21,7 +21,7 @@ public class Ejercicio10 {
 		System.out.println("Introduzca la longitud en metros");
 		
 		// Leemos los metros del teclado
-		metros = sc.nextFloat();
+		metros = sc.nextDouble();
 		
 		// Convertimos los metros a centimetros
 		centimetros = metros * 100;
